The Ladder Network is actually quite complex. Quite a bit more complex than the idea I have here.
Unrelated to that I will make a change to the way I do training.

I've mentioned this possibility, but the initial W should not be optimized using backpropagation. Instead I should carry the state.

And for training regular RNNs, I could do the same with regards to the state vector. It would be better to do it like that when doing RL.

...Ah, wait. That only applies when doing online learning. As soon as I step into minibatch learning that stops making sense.

Nevermind.