"10:20am. https://youtu.be/V2z80aXF8b4?list=PLnnr1O8OWc6bh5CYcqrAjfyzPH3QV745M&t=330
Oh, I did not know that Boltzman machines use the Hebbian learning rule. But here it is. There are many variants of it and this is definitely one of them.

11:05am. https://youtu.be/fzAuXMg_7n4?list=PLnnr1O8OWc6bh5CYcqrAjfyzPH3QV745M&t=71

There is no doubt that these videos are still difficult, but I think that unlike last time, I understand the essence of it. I understand sampling and I understand Hebbian learning. That is pretty much all that I need to understand Boltzman machines. Making reality stronger and fantasy weaker is really such a great learning rule. It is quite ingenious. I love it.

Come to think of it, this might be the reason why he mentioned that sleep exists to do unlearning. Dreams are fantasy."