Let me take a break here. I am really warming up to this new way of doing things. I think I'll start work on it when the time comes to resume. I am still enjoying myself by rolling it over in my mind.
Yeah, it is great. It has so much promise. This luxury of doing nothing while I just think about things is something I will avail myself to.

This idea will bring me that much closer to the Singularity. I do not think that brain does backprop, but I also think that backprop when done correctly will be a lot more powerful. For all the praise that the brain gets, I can see that it has some very serious limitations that backprop does not.

Once I cross the threshold I will just keep making greater and greater gains. Right now I am definitely making them, but it is not the case that those gains can be translated into actual force. I think that differentiable modulated plasticity will mark a change in that.