```         | .burn_in_state ->             inl {network} = states.last             Struct.foldr (inl layer _ ->                 match layer with                 | {init_state state} -> Struct.iter2 (inl to from -> s.CudaTensor.copy {from to}) init_state state                 | _ -> ()                 ) network () ```
I'll do it like this. I said nevermind to doing this with online learning, but it occurred to me that with online learning I can average the weights. This would essentially be the same thing Reptile does. And as init_state I can only set the weight. That means I do not need to update the vector. Brilliant, simply brilliant.

Let me modify the run function, the init and layer so it supports this.

Later I will add support for custom init_state update functions. For now let me just do this.