"7:30pm.
```
            inl fwd {input out m H} =
                H + n * (m * input * out - abs m * out * out * H)
            inl bck {input out m H} =
                {
                input = inl _ -> n * m * out
                out = inl _ -> n * (m * input - abs m * two * out * H)
                H = inl _ -> one - n * abs m * out * out
                m = inl _ -> n * (input * out - abs_bck m * out * out * H)
                }
```

```
inl abs_bck x = if x >= zero then one else -one
```

Yeah, it was this.

```
[|
    [|0.3682613; -0.0012082; -0.1091604; 0.1243474; -0.0519714; -0.3925951; 0.1847346; -0.8993623; -0.1700751; 0.2036941|]
|]

[|
    [|0.2797978; -0.008377166; -0.07566065; 0.1734103; -0.01168458; -0.009237467; -0.2485065; 0.2867839; -0.5718047; -0.222012|]
    [|0.2683595; -0.001114253; -0.01413604; -0.3275799; 0.007089609; -0.06014935; -0.1202063; -0.5583861; 0.05633133; -0.1370964|]
    [|-0.330192; -0.009641166; 0.07433189; 0.2131473; 0.01508966; 0.04614706; -0.1571491; 0.3735321; -0.3493008; -0.1352508|]
    [|0.3591245; 0.007693887; -0.008190138; 0.2807708; 0.01828428; 0.1006734; -0.08712708; 0.02750418; -0.05266025; 0.1125715|]
    [|0.5609795; 0.001592785; 0.03104587; -0.07362422; -0.01403356; -0.0215956; 0.3665734; 0.3792295; 0.4790412; 0.2447667|]
    [|-0.1869915; -0.01315663; 0.04678152; -0.0298894; -0.007672952; -0.07623366; -0.1766446; 0.04216999; 0.0943007; 0.2404741|]
    [|-0.489796; 0.003064564; -0.04861297; -0.2740277; -0.006726868; 0.07592315; -0.009943393; 0.4376051; 0.1760653; -0.2509285|]
    [|-0.08958758; -0.007242048; -0.05800781; -0.2715294; -0.003209037; 0.1074915; -0.19654; 0.3060991; 0.3486592; 0.1668072|]
    [|-0.209615; -0.008235892; 0.08056656; -0.1422843; 0.01414941; 0.05898963; 0.2241123; 0.1930788; 0.5353687; 0.2230486|]
    [|0.2398211; -0.009657101; 0.01721186; -0.1489571; -0.01302399; 0.04865741; -0.1899565; 0.389228; -0.1719713; -0.1731932|]
|]
```

Just like the other one, this rule is stable in terms of weight scaling.

Without a doubt, this is what I should be using. Well, either this or the tanh version maybe.

But this will allow it better control over learning speed.

```
The learning rate is 2 ** -10
At iteration 9 the cost is 10.0508284568787
At iteration 19 the cost is 9.98715963363648
At iteration 29 the cost is 9.95472888946533
At iteration 39 the cost is 9.89429473876953
At iteration 49 the cost is 10.0574407577515
At iteration 59 the cost is 9.83535470962524
At iteration 69 the cost is 9.86306324005127
At iteration 79 the cost is 9.96178016662598
At iteration 89 the cost is 9.80664768218994
At iteration 99 the cost is 9.93070878982544
At iteration 109 the cost is 9.86427240371704
```

The learning curve is like this.

I basically know why this is. The modulator simply does not have enough control over the updates to do what I want it to. With the way it is now, it is only a step away from being able to do unsupervised learning, but this problem is on the level of needing an actual program to deal with.

7:40pm. What I did today was only the beginning of the work. I admit to having fallen into wild delusions of what I was going to do with this update. The jagged cliffs of reality sure are sharp. But this is what it really means to be human.

On the plus side, I do not think I was wrong with the pieces I have here and the reasoning I had for wanting to put in modulation. Indeed, modulation is absolutely essential and it is a miracle that the network is capable of doing anything without it, let alone the binary pattern test. Without modulation Hebbian learning would simply be incapable of discrimination. The reason why it worked for Miconi is because the RNN somehow has the capacity to take advantage of it despite that.

What I am going to do tomorrow is build off what Miconi has spearheaded. I think the things I implemented in the last two days should be added to those Hebbian nets. I'll go from the top starting with the vanilla one. I'll make its plasticity coefficients dependent on the input and the previous output. Same goes for modulation.

While modulation might be necessary, it is simply impossible that this will work without more static infrastructure to serve as control.

Expressing control is key - the better the network is capable of channeling its own learning the more will it be capable of relying on its unsupervised component as opposed to its static parts.

8:05pm. I am still thinking about it. I think it might be worth it to add horizontal modulation as well.

8:10pm. Well, no forget it. The one I have now should be fine.

The binary pattern test is just so unusual. The classification tasks can afford to throw away information, but this one requires some intricate reasoning that I cannot grasp. One thing I am sure is that these metalearning nets would be great for recommender systems.

Let me stop here for the day. I think I'll just think about this while I lie in bed. Now that I've gotten a taste of torment, I'll go into it tomorrow with full force."