I did not get up just now. I remember it still being night when I came to my senses. I am still thinking about it obsessively.
But it does seem that I am making real progress in my understanding. I was wrong to think that plasticity in the DP paper has no control. I'd say that it has about as much control as modulation when `m` is passed through `abs`. The idea I had yesterday is not sound - I did not realize that the rule that I am using now actually leads to unlearning when the inputs are negative. This is unlike plasticity.

So on the problem...yeah, it should at least get half. Which it is not doing. But it might be simply due to the fact that learning the identity is in fact difficult.

I've been obsessively thinking, but I should do programming instead.

Also I think that at this juncture it is time that I start brining in dense connections. I had the idea of passing them through separate weight matrices to save memory, but that is not worth it. In feedforward layers, the inputs are not the main overhead. It is the computation and weights.

Let me start from the basics. First I need to make sure of various things.

Let me bring in a feedforward network and see if it can learn the identity. Then I will see if I can make the modulated feedforward net do the same.