"9:25am. Wow, it is really annoying how every net has too large weights for this shit.
```
    inl unmodulated_feedforward size =
        {
        init = inl sublayer_size ->
            {
            dsc =
                {
                input = Initializer.randn {dim=sublayer_size, size; stddev=0.01f32}
                bias = Initializer.bias size
                }
            size
            }

        apply = inl {weights input} -> matmultb (input, weights.input) weights.bias >>= Activation.tanh
        block = ()
        }
```

Let me do it like this in the Modulator segment.

9:25am. Yeah, this thing starts learning the identity immediatelly.

9:30am. Let me remove the state burn in.

9:30am.

```
The learning rate is 2 ** -10
At iteration 9 the cost is 10.5734022140503
At iteration 19 the cost is 9.91518306732178
At iteration 29 the cost is 9.21535768508911
At iteration 39 the cost is 9.14294447898865
At iteration 49 the cost is 9.18480410575867
At iteration 59 the cost is 9.65605797767639
At iteration 69 the cost is 8.42629165649414
At iteration 79 the cost is 10.0679758071899
At iteration 89 the cost is 7.91681571006775
At iteration 99 the cost is 9.10910677909851
At iteration 109 the cost is 7.11195154190063
At iteration 119 the cost is 8.02117338180542
At iteration 129 the cost is 8.56213068962097
At iteration 139 the cost is 6.64858059883118
At iteration 149 the cost is 6.21780548095703
At iteration 159 the cost is 7.66476292610168
At iteration 169 the cost is 6.65660500526428
At iteration 179 the cost is 5.97441174983978
At iteration 189 the cost is 6.0036750793457
At iteration 199 the cost is 4.65515329837799
At iteration 209 the cost is 3.73929773569107
At iteration 219 the cost is 3.83242356777191
At iteration 229 the cost is 2.74873661994934
At iteration 239 the cost is 3.52706569433212
At iteration 249 the cost is 3.52700377702713
At iteration 259 the cost is 2.55626110434532
At iteration 269 the cost is 2.3555585205555
At iteration 279 the cost is 2.98712728023529
At iteration 289 the cost is 2.13939062356949
At iteration 299 the cost is 2.22175006121397
At iteration 309 the cost is 1.75728176832199
At iteration 319 the cost is 1.81165431961417
At iteration 329 the cost is 1.18037874847651
At iteration 339 the cost is 1.43056602776051
At iteration 349 the cost is 1.58810597658157
At iteration 359 the cost is 1.59617658257484
At iteration 369 the cost is 0.653564175218344
At iteration 379 the cost is 1.43616574704647
At iteration 389 the cost is 1.09780423045158
At iteration 399 the cost is 0.889806383103132
At iteration 409 the cost is 1.56294990181923
At iteration 419 the cost is 0.610232338681817
At iteration 429 the cost is 0.845834086462855
At iteration 439 the cost is 0.579389877803624
At iteration 449 the cost is 0.622540828026831
At iteration 459 the cost is 0.423470686003566
At iteration 469 the cost is 0.566224949806929
At iteration 479 the cost is 0.238024072349072
At iteration 489 the cost is 0.316543398052454
At iteration 499 the cost is 0.343497842922807
```

Hmmm, what the fuck? I just set the `n` to 1f32.

But the net should not be capable of doing this. I spent the whole night coming to this conclusion."