Here are my thoughts. I thought a lot about this during the night.
Yesterday I wrote that the standard plastic RNN has no control and that it is amazing that it manages to do anything at all. But that is not correct.

If I think about what sort of weights the process of Hebbian learning is most likely to converge to then the obvious candidate would be the bias term. The result of Hebbian learning would probabilistically depend on the bias weights. Originally this is what I wanted modulation for, and indeed, modulation should result in moving the learning rate towards the template when it is positive.

It might be worth it to use modulation in order to inhibit learning rather than do something the bias weights already do.

...This thing right now works just horribly. Is the implementation even correct? Agh, had I been using an AD library this would never be in question. Trully troublesome. Spiral is not the best fit for research.

Let me try initing the input weights to the identity.