"1:50pm. I know what I should be trying to do. `(one - relu m)`. Let me try this sort of activation for `m`.
1:55pm.

```
            inl fwd {input out m H} =
                H + n * (one - relu_fwd m) * (input * out - out * out * H)
            inl bck {input out m H} =
                {
                input = inl _ -> (one - relu_fwd m) * n * out
                out = inl _ -> (one - relu_fwd m) * n * (input - two * out * H)
                H = inl _ -> one - (one - relu_fwd m) * n * out * out
                m = inl _ -> -(relu_bck m) * n * (input * out - out * out * H)
                }
```

This seems to be working to a degree at least.

1:55pm. I can't accept that tanh is working poorly initialy. Most likely that I need to do is set its bias to 1.

Also I want to stop doing the activation for m inside this function. Maybe once I've settled on one, but this is near the limit where I am starting to have difficulties keeping track of things. What was the thing where `abs m` was on the right side.

2pm.

```
            inl fwd {input out m H} =
                H + n * (m * input * out - abs m * out * out * H)
            inl bck {input out m H} =
                {
                input = inl _ -> n * m * out
                out = inl _ -> n * (m * input - abs m * two * out * H)
                H = inl _ -> one - n * abs m * out * out
                m = inl _ -> n * (input * out - abs_bck m * out * out * H)
                }
```

It was this thing.

Let me give it a shot and then I'll set the biases to 1 to start things off."