"3:45pm. But in all likelihood, I am still probably initializing the modulator too high.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.4732809066772
At iteration 19 the cost is 23.1491306304932
At iteration 29 the cost is 20.4202949523926
At iteration 39 the cost is 22.5748213768005
At iteration 49 the cost is 18.4376484394073
At iteration 59 the cost is 15.0010992050171
At iteration 69 the cost is 16.9917044639587
At iteration 79 the cost is 19.5236093521118
At iteration 89 the cost is 14.7476576805115
At iteration 99 the cost is 12.2967095971107
At iteration 109 the cost is 14.9617751598358
At iteration 119 the cost is 10.3415517807007
At iteration 129 the cost is 12.9289459705353
At iteration 139 the cost is 10.4377179861069
At iteration 149 the cost is 7.59788553714752
At iteration 159 the cost is 9.18797402381897
At iteration 169 the cost is 9.28644161224365
At iteration 179 the cost is 8.7677830696106
At iteration 189 the cost is 6.32587630748749
At iteration 199 the cost is 6.50227627754211
At iteration 209 the cost is 7.44228035211563
At iteration 219 the cost is 5.11135244965553
At iteration 229 the cost is 4.85083322525024
At iteration 239 the cost is 4.68875237107277
At iteration 249 the cost is 7.369287882559
At iteration 259 the cost is 3.48799127223901
At iteration 269 the cost is 7.71015848517418
At iteration 279 the cost is 7.57650072574616
At iteration 289 the cost is 7.93322006464004
At iteration 299 the cost is 4.89419829845428
At iteration 309 the cost is 7.67354613505304
At iteration 319 the cost is 6.69558448791504
At iteration 329 the cost is 4.706254529953
At iteration 339 the cost is 4.50973527505994
At iteration 349 the cost is 5.65484795570374
At iteration 359 the cost is 7.30082168579102
At iteration 369 the cost is 7.2871640920639
At iteration 379 the cost is 5.93691681623459
At iteration 389 the cost is 3.48757489919662
At iteration 399 the cost is 5.84175870418549
At iteration 409 the cost is 5.39284541606903
At iteration 419 the cost is 5.80586410164833
At iteration 429 the cost is 9.98344883918762
At iteration 439 the cost is 5.58368272781372
At iteration 449 the cost is 6.27931022644043
At iteration 459 the cost is 4.27353864908218
At iteration 469 the cost is 7.5830967605114
At iteration 479 the cost is 2.88143294453621
At iteration 489 the cost is 3.00546572357416
At iteration 499 the cost is 6.32198283076286
At iteration 509 the cost is 6.07741631269455
At iteration 519 the cost is 6.77673243284225
At iteration 529 the cost is 4.76933535337448
At iteration 539 the cost is 5.99635894298553
At iteration 549 the cost is 9.57996808290482
At iteration 559 the cost is 6.74599540233612
At iteration 569 the cost is 6.15293563604355
At iteration 579 the cost is 4.48055462390184
At iteration 589 the cost is 6.0474916934967
At iteration 599 the cost is 5.78187556862831
At iteration 609 the cost is 4.66986107826233
At iteration 619 the cost is 5.87321972846985
At iteration 629 the cost is 3.94900671839714
At iteration 639 the cost is 6.98107948303223
At iteration 649 the cost is 12.3776618242264
At iteration 659 the cost is 6.4958263874054
At iteration 669 the cost is 5.79342064857483
At iteration 679 the cost is 7.88893322944641
At iteration 689 the cost is 6.35144574642181
At iteration 699 the cost is 6.57498467564583
At iteration 709 the cost is 3.92650744915009
At iteration 719 the cost is 7.55221415758133
At iteration 729 the cost is 8.90043489933014
At iteration 739 the cost is 5.57051348686218
At iteration 749 the cost is 5.16331605315208
At iteration 759 the cost is 5.94916210174561
At iteration 769 the cost is 10.0972036004066
At iteration 779 the cost is 6.33372406959534
At iteration 789 the cost is 7.19027149677277
At iteration 799 the cost is 7.58476195335388
At iteration 809 the cost is 9.03326237201691
At iteration 819 the cost is 7.21926891803741
At iteration 829 the cost is 7.64394557476044
At iteration 839 the cost is 6.88662343025208
At iteration 849 the cost is 10.8624255895615
At iteration 859 the cost is 8.74483790397644
At iteration 869 the cost is 8.10451488494873
At iteration 879 the cost is 4.9662491440773
At iteration 889 the cost is 8.49574179649353
At iteration 899 the cost is 8.04503116607666
At iteration 909 the cost is 10.824812734127
At iteration 919 the cost is 10.9777539968491
At iteration 929 the cost is 10.1279203295708
At iteration 939 the cost is 9.84137077331543
At iteration 949 the cost is 10.7177502393723
At iteration 959 the cost is 11.1733504891396
At iteration 969 the cost is 8.56849136352539
At iteration 979 the cost is 11.5928713321686
At iteration 989 the cost is 6.66799728870392
At iteration 999 the cost is 13.1878355503082
At iteration 1009 the cost is 11.3052849292755
At iteration 1019 the cost is 9.10695233345032
At iteration 1029 the cost is 7.47431383132935
At iteration 1039 the cost is 10.2368571519852
At iteration 1049 the cost is 10.7369010448456
At iteration 1059 the cost is 12.0996057987213
```

Yeah, I was initializing it too high. It does really, really well at the beginning like this. In fact, maybe I could get the same effect by setting the alpha higher in the regular oja. Let me give it a try.

4pm. No, unlike what I expected, the acceleration from making plasticity dependent on the inputs seems to be genuine. Score!

This is exactly what I was looking for to improve the performance of these algorithms.

Let me try initing the modulators lower once more."