"4:05pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.5213705062866
At iteration 19 the cost is 22.9692378044128
At iteration 29 the cost is 20.3596080780029
At iteration 39 the cost is 22.5558381080627
At iteration 49 the cost is 18.6283213615417
At iteration 59 the cost is 14.9500749588013
At iteration 69 the cost is 17.2496254920959
At iteration 79 the cost is 19.5559001922607
At iteration 89 the cost is 14.7508772850037
At iteration 99 the cost is 12.3056188106537
At iteration 109 the cost is 14.9378086090088
At iteration 119 the cost is 9.9616955280304
At iteration 129 the cost is 12.7812068939209
At iteration 139 the cost is 10.4664397954941
At iteration 149 the cost is 7.46700644493103
At iteration 159 the cost is 9.10920240879059
At iteration 169 the cost is 9.24154376983643
At iteration 179 the cost is 8.71550915241241
At iteration 189 the cost is 6.52698982954025
At iteration 199 the cost is 6.29219448566437
At iteration 209 the cost is 7.49909424185753
At iteration 219 the cost is 5.11906495988369
At iteration 229 the cost is 4.9571485966444
At iteration 239 the cost is 4.60960215330124
At iteration 249 the cost is 7.17516853939742
At iteration 259 the cost is 3.67583807264455
At iteration 269 the cost is 7.71508636474609
At iteration 279 the cost is 7.15231640338898
At iteration 289 the cost is 7.54830280542374
At iteration 299 the cost is 4.64833180904388
At iteration 309 the cost is 7.56021207720041
At iteration 319 the cost is 6.45359598398209
At iteration 329 the cost is 4.34157895445824
At iteration 339 the cost is 4.07263212278485
At iteration 349 the cost is 5.712575507164
At iteration 359 the cost is 6.32981151342392
At iteration 369 the cost is 6.18902026414871
At iteration 379 the cost is 5.9580299705267
At iteration 389 the cost is 3.26786688566208
At iteration 399 the cost is 5.48085286617279
At iteration 409 the cost is 5.03712701797485
```

This is with a really low modulator init. I am starting to wonder if after a certain point that even means anything. Let me init it to zero.

4:05pm.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.5702268600464
At iteration 19 the cost is 22.968772315979
At iteration 29 the cost is 19.9075162887573
At iteration 39 the cost is 22.380283164978
At iteration 49 the cost is 18.62991335392
At iteration 59 the cost is 15.1309515476227
At iteration 69 the cost is 17.6246955871582
At iteration 79 the cost is 19.6113586425781
At iteration 89 the cost is 15.2703279018402
At iteration 99 the cost is 12.1846227645874
At iteration 109 the cost is 15.0255967140198
At iteration 119 the cost is 10.0145649909973
At iteration 129 the cost is 12.7410594940186
At iteration 139 the cost is 10.5121950149536
At iteration 149 the cost is 7.41971597671509
At iteration 159 the cost is 9.17069489955902
At iteration 169 the cost is 9.09259485006332
At iteration 179 the cost is 8.1356240272522
At iteration 189 the cost is 6.96914554834366
At iteration 199 the cost is 6.44951756000519
At iteration 209 the cost is 7.89727185964584
```

Yeah, nice. Quite nice. This is with both the bias and the weight for the modulators set to zero initially.

Let me try lowering the learning rate to see if that stabilizes things a little. Right now it is a bit too fast."