"4:50pm.
```
                        generalized_mi {input state bias=weights.bias}
```

Implementing MI is really, really easy.

Let me give it a shot.

4:55pm. It is a bit slower. Let me lower the init and increase the learning rate.

5pm.

```
The learning rate is 2 ** -10
At iteration 9 the cost is 30.8772733688355
At iteration 19 the cost is 27.0344755172729
At iteration 29 the cost is 22.5778596878052
At iteration 39 the cost is 20.6970430374146
At iteration 49 the cost is 15.5894391059875
At iteration 59 the cost is 12.7284600734711
At iteration 69 the cost is 11.8071005821228
At iteration 79 the cost is 11.9916282176971
At iteration 89 the cost is 10.6267909049988
At iteration 99 the cost is 8.99420275688171
At iteration 109 the cost is 11.7717483520508
At iteration 119 the cost is 7.80569524765015
At iteration 129 the cost is 9.14747934341431
At iteration 139 the cost is 9.09941539764404
At iteration 149 the cost is 7.5023353099823
At iteration 159 the cost is 7.50528547763824
At iteration 169 the cost is 8.73513066768646
At iteration 179 the cost is 8.93095126152039
At iteration 189 the cost is 8.85612788200378
At iteration 199 the cost is 7.85854804515839
At iteration 209 the cost is 7.02629871368408
At iteration 219 the cost is 6.06766648292542
At iteration 229 the cost is 6.63732588291168
At iteration 239 the cost is 7.23962297439575
At iteration 249 the cost is 7.57044486999512
At iteration 259 the cost is 6.83150079250336
At iteration 269 the cost is 8.97001037597656
At iteration 279 the cost is 7.44443001747131
At iteration 289 the cost is 7.46666967868805
At iteration 299 the cost is 6.99154829978943
At iteration 309 the cost is 9.77542006969452
At iteration 319 the cost is 7.09446775913239
At iteration 329 the cost is 8.13191452026367
At iteration 339 the cost is 6.58019814491272
At iteration 349 the cost is 8.27757167816162
At iteration 359 the cost is 7.24490995407105
At iteration 369 the cost is 7.54757933616638
At iteration 379 the cost is 8.88154273033142
At iteration 389 the cost is 7.64909725189209
At iteration 399 the cost is 6.63829388618469
At iteration 409 the cost is 7.09893646240234
At iteration 419 the cost is 10.593354177475
At iteration 429 the cost is 10.6037770748138
At iteration 439 the cost is 8.32776198387146
At iteration 449 the cost is 8.47824511528015
At iteration 459 the cost is 8.68055927753448
At iteration 469 the cost is 8.3174825668335
At iteration 479 the cost is 7.45019679069519
At iteration 489 the cost is 9.69718074798584
At iteration 499 the cost is 9.59497776031494
At iteration 509 the cost is 8.44136724472046
At iteration 519 the cost is 8.59977431297302
At iteration 529 the cost is 8.3907365322113
At iteration 539 the cost is 7.53169460296631
At iteration 549 the cost is 10.5109316825867
At iteration 559 the cost is 11.2764713525772
At iteration 569 the cost is 11.4818859100342
At iteration 579 the cost is 12.2945260047913
At iteration 589 the cost is 11.729080581665
At iteration 599 the cost is 11.5758748054504
At iteration 609 the cost is 9.68653788566589
At iteration 619 the cost is 9.29008893966675
At iteration 629 the cost is 12.2211657524109
At iteration 639 the cost is 9.7148407459259
At iteration 649 the cost is 13.6514675140381
At iteration 659 the cost is 11.2437606811523
At iteration 669 the cost is 11.2126183509827
At iteration 679 the cost is 14.612114906311
```

No, it seems that mi does not work well with this task. Let me lower the learning rate what it was. Maybe it will have better asymptotic performance, but I doubt it. I think it is fine even if I use a vanilla RNN here given that is far from what it started from. I am using MI on the plasticity vector and that is what really matters."