"5:05pm.
```
The learning rate is 2 ** -11.5
At iteration 9 the cost is 31.7841495513916
At iteration 19 the cost is 30.5193851470947
At iteration 29 the cost is 29.0164323806763
At iteration 39 the cost is 28.4964265823364
At iteration 49 the cost is 26.6525625228882
At iteration 59 the cost is 24.9735116958618
At iteration 69 the cost is 23.8720935821533
At iteration 79 the cost is 23.1557605743408
At iteration 89 the cost is 20.5892347335815
At iteration 99 the cost is 19.3133662223816
At iteration 109 the cost is 19.0820280075073
At iteration 119 the cost is 16.2313854217529
At iteration 129 the cost is 15.5119275093079
At iteration 139 the cost is 14.4128887176514
At iteration 149 the cost is 13.2659135818481
At iteration 159 the cost is 13.0641224861145
At iteration 169 the cost is 12.4148164749146
At iteration 179 the cost is 12.3171545982361
At iteration 189 the cost is 12.0297513008118
At iteration 199 the cost is 11.1300971031189
At iteration 209 the cost is 11.2590016841888
At iteration 219 the cost is 10.5531225204468
At iteration 229 the cost is 9.5579225063324
At iteration 239 the cost is 9.53883118629456
At iteration 249 the cost is 9.97457337379456
At iteration 259 the cost is 9.52475347518921
At iteration 269 the cost is 10.4612154960632
At iteration 279 the cost is 9.79202489852905
At iteration 289 the cost is 8.43093481063843
At iteration 299 the cost is 8.08912353515625
At iteration 309 the cost is 9.51350636482239
At iteration 319 the cost is 7.40761713981628
At iteration 329 the cost is 7.65311920642853
At iteration 339 the cost is 7.12914786338806
At iteration 349 the cost is 7.83329520225525
At iteration 359 the cost is 8.03724675178528
At iteration 369 the cost is 7.3142561674118
At iteration 379 the cost is 6.57840223312378
At iteration 389 the cost is 6.73629894256592
At iteration 399 the cost is 6.21158883571625
At iteration 409 the cost is 5.09725842475891
At iteration 419 the cost is 7.83119316101074
At iteration 429 the cost is 8.43251979351044
At iteration 439 the cost is 6.15904378890991
At iteration 449 the cost is 6.72458500862122
At iteration 459 the cost is 5.48407251834869
At iteration 469 the cost is 6.36801114082336
At iteration 479 the cost is 5.37616651058197
At iteration 489 the cost is 5.62194437980652
At iteration 499 the cost is 5.5643966794014
At iteration 509 the cost is 5.00313605666161
At iteration 519 the cost is 4.33663449287415
At iteration 529 the cost is 5.23098553419113
At iteration 539 the cost is 4.29998602867127
At iteration 549 the cost is 7.02541625499725
At iteration 559 the cost is 6.30069420337677
At iteration 569 the cost is 5.12115740776062
At iteration 579 the cost is 5.1230898976326
At iteration 589 the cost is 5.0267566204071
At iteration 599 the cost is 4.78205773830414
At iteration 609 the cost is 4.74765951633453
At iteration 619 the cost is 4.56072881221771
At iteration 629 the cost is 3.47762569189072
At iteration 639 the cost is 5.47193028926849
At iteration 649 the cost is 6.75059292316437
At iteration 659 the cost is 5.18830373287201
At iteration 669 the cost is 4.44258130788803
At iteration 679 the cost is 3.67397826910019
At iteration 689 the cost is 4.21173567771912
At iteration 699 the cost is 3.90070540904999
At iteration 709 the cost is 4.05240911245346
At iteration 719 the cost is 5.45039173364639
At iteration 729 the cost is 4.69001004695892
At iteration 739 the cost is 4.08739150762558
At iteration 749 the cost is 3.08027129173279
At iteration 759 the cost is 3.40497769713402
At iteration 769 the cost is 4.56973212957382
At iteration 779 the cost is 5.39413702487946
At iteration 789 the cost is 3.87675268054008
At iteration 799 the cost is 5.99880380630493
At iteration 809 the cost is 5.67635872364044
At iteration 819 the cost is 4.68597921133041
At iteration 829 the cost is 3.80284161567688
At iteration 839 the cost is 4.3154461145401
At iteration 849 the cost is 5.06244881153107
At iteration 859 the cost is 4.94430873394012
At iteration 869 the cost is 5.57379233837128
At iteration 879 the cost is 3.74810838103294
At iteration 889 the cost is 4.05583565235138
At iteration 899 the cost is 4.61501599550247
At iteration 909 the cost is 6.12965111136436
At iteration 919 the cost is 3.66574972867966
At iteration 929 the cost is 7.07339198589325
At iteration 939 the cost is 3.37093138098717
At iteration 949 the cost is 7.56915650367737
At iteration 959 the cost is 5.29660725593567
At iteration 969 the cost is 5.48732817173004
At iteration 979 the cost is 5.1247126147151
At iteration 989 the cost is 3.48798408508301
At iteration 999 the cost is 4.61580581665039
At iteration 1009 the cost is 6.58220686912537
At iteration 1019 the cost is 4.17929371595383
At iteration 1029 the cost is 3.68979130983353
At iteration 1039 the cost is 4.32853090763092
At iteration 1049 the cost is 6.21976482868195
At iteration 1059 the cost is 4.19932565689087
At iteration 1069 the cost is 3.06974102258682
At iteration 1079 the cost is 4.19028999507427
At iteration 1089 the cost is 6.69499475955963
At iteration 1099 the cost is 3.68787753582001
At iteration 1109 the cost is 4.14862756729126
At iteration 1119 the cost is 5.19255269765854
At iteration 1129 the cost is 5.6503578543663
At iteration 1139 the cost is 5.55116546154022
At iteration 1149 the cost is 3.99166066646576
At iteration 1159 the cost is 5.22399590015411
At iteration 1169 the cost is 3.96142338514328
At iteration 1179 the cost is 4.39003286361694
At iteration 1189 the cost is 4.92729405164719
At iteration 1199 the cost is 6.38802833557129
At iteration 1209 the cost is 8.13922489881515
At iteration 1219 the cost is 4.40676383972168
At iteration 1229 the cost is 4.18165447711945
At iteration 1239 the cost is 5.01709933280945
At iteration 1249 the cost is 5.85563163757324
At iteration 1259 the cost is 4.27336719036102
At iteration 1269 the cost is 4.61587965488434
At iteration 1279 the cost is 6.56605751514435
At iteration 1289 the cost is 5.89688405990601
At iteration 1299 the cost is 5.37319690585136
At iteration 1309 the cost is 8.70746073722839
At iteration 1319 the cost is 5.79272048473358
At iteration 1329 the cost is 4.51121701002121
```

It is not the full run, but I ran out of patience. MI on the input and the state is useless here. It is just getting in the way.

5:05pm. Let me run the vanilla RNN one with a lower learning rate just so I can see if it can go to below 1 consistently and then I will call it a day. This performance is quite satisfactory. I might look into combining it with LSTMs tomorrow or something like that. The discovery is unexpected so I haven't digested it yet properly."