"8:30am.
```
The learning rate is 2 ** -13.5
At iteration 9 the cost is 25.958978843689
At iteration 19 the cost is 24.5500858306885
At iteration 29 the cost is 23.1138113021851
At iteration 39 the cost is 23.0693082809448
At iteration 49 the cost is 20.4915301322937
At iteration 59 the cost is 17.0992058753967
At iteration 69 the cost is 18.7218252182007
At iteration 79 the cost is 16.9439660072327
At iteration 89 the cost is 13.8405766010284
At iteration 99 the cost is 12.7345090150833
At iteration 109 the cost is 14.5863082408905
At iteration 119 the cost is 9.40935859680176
At iteration 129 the cost is 11.3516245126724
At iteration 139 the cost is 12.1467974185944
At iteration 149 the cost is 9.220901876688
At iteration 159 the cost is 14.3597661972046
At iteration 169 the cost is 13.482293844223
At iteration 179 the cost is 13.1699967861176
At iteration 189 the cost is 11.230518913269
At iteration 199 the cost is 11.3149103403091
At iteration 209 the cost is 12.3515376567841
At iteration 219 the cost is 8.81043110489845
At iteration 229 the cost is 11.3655891656876
At iteration 239 the cost is 8.49541742801666
At iteration 249 the cost is 13.7001699090004
At iteration 259 the cost is 8.20861109495163
At iteration 269 the cost is 12.7354199886322
At iteration 279 the cost is 13.8455555200577
At iteration 289 the cost is 10.8813096523285
At iteration 299 the cost is 11.7457145214081
At iteration 309 the cost is 13.8304881572723
At iteration 319 the cost is 11.9545331954956
At iteration 329 the cost is 11.6547597169876
At iteration 339 the cost is 12.6059675931931
At iteration 349 the cost is 16.5140058040619
At iteration 359 the cost is 19.7015542030334
At iteration 369 the cost is 12.7931904315948
At iteration 379 the cost is 14.7263641357422
At iteration 389 the cost is 13.297208404541
At iteration 399 the cost is 12.265592956543
At iteration 409 the cost is 16.7890230894089
At iteration 419 the cost is 15.5263729572296
At iteration 429 the cost is 17.177411365509
At iteration 439 the cost is 12.2943971633911
At iteration 449 the cost is 16.357075548172
At iteration 459 the cost is 13.1070966720581
At iteration 469 the cost is 15.3584482192993
At iteration 479 the cost is 11.3334821939468
At iteration 489 the cost is 11.8868785858154
At iteration 499 the cost is 16.7442447662354
At iteration 509 the cost is 17.0070161819458
At iteration 519 the cost is 13.5318580627441
At iteration 529 the cost is 21.5978753089905
At iteration 539 the cost is 98.5538856506348
At iteration 549 the cost is 101.427612304688
At iteration 559 the cost is 101.94815826416
At iteration 569 the cost is 97.2317169189453
```

It is unstable. Let me try removing that abs n thing, though I do not think this would be the problem.

8:35am.

```
The learning rate is 2 ** -13.5
At iteration 9 the cost is 25.9830303192139
At iteration 19 the cost is 24.7669538497925
At iteration 29 the cost is 23.6203609466553
At iteration 39 the cost is 23.6657932281494
At iteration 49 the cost is 22.0216623306274
At iteration 59 the cost is 20.674073600769
At iteration 69 the cost is 20.3783912658691
At iteration 79 the cost is 20.326812171936
At iteration 89 the cost is 18.0981533050537
At iteration 99 the cost is 17.7612756729126
At iteration 109 the cost is 17.4886637687683
At iteration 119 the cost is 15.3556854248047
At iteration 129 the cost is 15.9697823524475
At iteration 139 the cost is 14.7943048477173
At iteration 149 the cost is 12.9629199981689
At iteration 159 the cost is 13.7146348953247
At iteration 169 the cost is 12.1300902366638
At iteration 179 the cost is 12.1701949596405
At iteration 189 the cost is 10.8300765514374
At iteration 199 the cost is 9.66194458007812
At iteration 209 the cost is 11.0831166267395
At iteration 219 the cost is 8.21150524616241
At iteration 229 the cost is 7.92162256240845
At iteration 239 the cost is 7.68749341964722
At iteration 249 the cost is 8.99640221595764
At iteration 259 the cost is 6.14699611663818
At iteration 269 the cost is 9.01703128814697
At iteration 279 the cost is 8.34198265075684
At iteration 289 the cost is 6.66072201728821
At iteration 299 the cost is 6.02986613512039
At iteration 309 the cost is 7.75108177661896
At iteration 319 the cost is 5.4774486541748
At iteration 329 the cost is 5.64239141941071
At iteration 339 the cost is 4.60567736625671
At iteration 349 the cost is 5.15803495645523
At iteration 359 the cost is 8.06357760429382
At iteration 369 the cost is 6.49539716243744
At iteration 379 the cost is 5.99331004619598
At iteration 389 the cost is 5.10482822060585
At iteration 399 the cost is 4.41889779567718
At iteration 409 the cost is 5.04582768082619
At iteration 419 the cost is 5.24777463674545
At iteration 429 the cost is 7.22408201694489
At iteration 439 the cost is 5.50504460334778
At iteration 449 the cost is 5.1024015724659
At iteration 459 the cost is 3.75911762714386
At iteration 469 the cost is 6.12179098129272
At iteration 479 the cost is 2.96466902494431
At iteration 489 the cost is 4.82666143774986
At iteration 499 the cost is 4.38610035181046
At iteration 509 the cost is 4.54520934820175
At iteration 519 the cost is 2.78724300265312
At iteration 529 the cost is 5.30982377529144
At iteration 539 the cost is 2.37051974534988
At iteration 549 the cost is 7.80851318836212
At iteration 559 the cost is 5.78496226668358
At iteration 569 the cost is 4.67161948680878
At iteration 579 the cost is 2.43830388486385
At iteration 589 the cost is 6.08739700317383
At iteration 599 the cost is 6.02981590032578
At iteration 609 the cost is 3.17055694460869
At iteration 619 the cost is 4.49823708534241
At iteration 629 the cost is 6.08073702454567
At iteration 639 the cost is 6.27738537788391
At iteration 649 the cost is 5.41690617799759
At iteration 659 the cost is 7.1230825304985
At iteration 669 the cost is 4.2149994134903
At iteration 679 the cost is 4.77710857689381
At iteration 689 the cost is 4.26390736103058
At iteration 699 the cost is 2.25166377425194
At iteration 709 the cost is 5.31166576743126
```

No, it seems that making `n` state-input dependent is a poor choice. And it seem the original Oja's rule is more stable. I will have to backtrack in my later mail.

...Let me make `n` the standard RNN formulation. Then if that fails I'll try making it a regular bias term. If that also fails I'll go back to using it as a hyperparameter."